Former-commit-id: 289aef7d38b196bf75efe948bfd0d26bd1c0c1f3

+++ back/Spring/fisher/src/main/java/com/net/fisher/post/service/PostService.java
@@ -158,15 +158,17 @@ public class PostService {
         }
 
         // 이미지 삭제 할 것 있으면 삭제
-        for(long id : postPatchDto.getDeleteImageList()){
-            deleteImage(id);
-        }
-    }
+        if(postPatchDto.getDeleteImageList() != null){
+            for(long id : postPatchDto.getDeleteImageList()){
+                PostImage postImage = postImageRepository.findById(id).orElseThrow(() -> new BusinessLogicException(ExceptionCode.FILE_NOT_FOUND));
 
-    @Transactional
-    public void deleteImage(long fileId) {
-        PostImage postImage = postImageRepository.findById(fileId).orElseThrow(() -> new BusinessLogicException(ExceptionCode.FILE_NOT_FOUND));
-        postImageRepository.delete(postImage);
+                if(postImage.getPost() != post){
+                    throw new BusinessLogicException(ExceptionCode.FAILED_TO_UPDATE_FILE);
+                }
+
+                postImageRepository.delete(postImage);
+            }
+        }
     }
 
     @Transactional
